third_party/blink/web_tests/fast/frames/frame-dead-region-expected.txt

ALERT: SUCCESS -- the link was clicked
